Add JWT auth to attachment and user search routes
- Extract getAuthUserId to helpers.ts for reuse across route modules.
- Apply JWT auth to all attachment endpoints (upload, download, delete).
- Require JWT auth for user search endpoint.
- Update tests to pass Bearer tokens for attachment and user routes.
